backend/requirements.txt
fastapi~=0.87.0
uvicorn~=0.20.0
pillow~=9.3.0
rembg==2.0.30
